New file: Deluxe Stand for Maslow CNC
This is a Sketchup file of a deluxe stand for the Maslow CNC machine.  Features include: 1) a 3/4" sheet of plywood at 15 degrees from vertical, just like to original Maslow CNC stand; 2) locking casters for ease of movement; 3) space for plywood storage; 4) a whiteboard; and 5) rails for a panel saw.  In the current version, I have moved the CNC base outward 1.5" to avoid the risk of the router striking the panel saw rails, per a warning from the Forums.